[#1590] FE: move modal & members component to modules
